Making more generalized to fit other image uses
By changing this to raster instead of base you can now input images that
DO NOT have the same number of bands and resize the images spatially.
So input or base can have 224, but other images can have fewer. This
WILL NOT result in images the same size unless the base image and
resized images have same number of bands.